Combined ssd300_blocks and ssd512_block for better overview.
