+++ ssd/ssd_methods.py
@@ -144,17 +144,14 @@ def detected_bboxes(predictions, localisations, num_classes,
     """Get the detected bounding boxes from the SSD network output.
     """
     # Select top_k bboxes from predictions, and clip
-    rscores, rbboxes = \
-        ssd_common.tf_ssd_bboxes_select(predictions, localisations,
-                                        select_threshold=select_threshold,
-                                        num_classes=num_classes)
-    rscores, rbboxes = \
-        tfe.bboxes_sort(rscores, rbboxes, top_k=top_k)
+    rscores, rbboxes = ssd_common.tf_ssd_bboxes_select(predictions, localisations,
+                                                       select_threshold=select_threshold,
+                                                       num_classes=num_classes)
+    rscores, rbboxes = tfe.bboxes_sort(rscores, rbboxes, top_k=top_k)
     # Apply NMS algorithm.
-    rscores, rbboxes = \
-        tfe.bboxes_nms_batch(rscores, rbboxes,
-                             nms_threshold=nms_threshold,
-                             keep_top_k=keep_top_k)
+    rscores, rbboxes = tfe.bboxes_nms_batch(rscores, rbboxes,
+                                            nms_threshold=nms_threshold,
+                                            keep_top_k=keep_top_k)
     if clipping_bbox is not None:
         rbboxes = tfe.bboxes_clip(clipping_bbox, rbboxes)
     return rscores, rbboxes
